WIP: Store UA stylesheets on the StyleComputer
Without this, we keep the UA stylesheets forever. This means they keep a
reference to whichever Realm was active when they were first created,
thus leaking that Realm and whatever it references.

This attempt makes the stylesheets members of the StyleComputer, so
they're created with the document, and are freed when the document is.
The downside is that this means we have multiple copies in memory at
once, especially on pages with a lot of iframes and such. With testing,
this doesn't seem to save any memory - in fact we were using more,
though I don't know how accurate my testing was.

All of this is only a temporary measure anyway, until we remove the
CSSOM requirement from our internal types, so it doesn't seem worth it.