LibWeb: Remove now-invalid attr() type support
Previously the type argument in attr() could be the name of a CSS type
on its own. This has changed, and now only `string` or the name of a
dimension unit is allowed. Other types and more complex grammar use the
`type()` function, which we don't yet support.

I've updated the syntax comment, but not the algorithm itself, which
will be reimplemented in a later commit.