WIP: LibWeb/CSS: Add PendingSubstitutionStyleValue
The idea is that any shorthands with ArbSubFuns in can't be expanded
into their longhands until computed-value time, but still need to
cascade those longhands. In that case, a PendingSubstitutionStyleValue
will be cascaded instead, and then once the shorthand is resolved, it
can replace these pending values with the real ones.

This can actually be a separate PR to be honest.

Still needs doing: replacing *only* the PSSVs after resolving the
shorthand. I think currently we replace all the longhands? I can't find
where that's happening though.