WIP: Descriptor serialization
What I started working on last here, is having a centralized "make a
shorthand out of these longhands" function in Shorthands.{h,cpp}. It'll
be a little messy because of having Properties and Descriptors, but ehh.

Anyway it's a huge tangent from `@page` so it's going on a branch for
later.